Combined Console and TextConsole
