+++ src/com/jhjavadev/textengine/game/Stage.java
@@ -1,14 +1,14 @@
 package com.jhjavadev.textengine.game;
 
-import com.jhjavadev.textengine.console.TextConsole;
+import com.jhjavadev.textengine.console.Console;
 
 public abstract class Stage {
 	protected int pos = 0;
 	protected boolean done = false;
 
-	public abstract void printStart(TextConsole c);
-	public abstract boolean update(String cmd, String[] args, TextConsole c);
-	public abstract void printHelp(TextConsole c);
+	public abstract void printStart(Console c);
+	public abstract boolean update(String cmd, String[] args, Console c);
+	public abstract void printHelp(Console c);
 	public abstract void reset();
 
 	public boolean isDone() {
