+++ src/com/jhjavadev/textengine/game/TestStage.java
@@ -1,18 +1,18 @@
 package com.jhjavadev.textengine.game;
 
-import com.jhjavadev.textengine.console.TextConsole;
+import com.jhjavadev.textengine.console.Console;
 
 public class TestStage extends Stage {
 
-	public void printStart(TextConsole c) {
+	public void printStart(Console c) {
 
 	}
 
-	public boolean update(String cmd, String[] args, TextConsole c) {
+	public boolean update(String cmd, String[] args, Console c) {
 		return false;
 	}
 
-	public void printHelp(TextConsole c) {
+	public void printHelp(Console c) {
 
 	}
